% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soilCores}
\alias{soilCores}
\title{Example unstandardized soil core dataset}
\format{
a dataframe containing the necessary columns for the soil_standardize function to work.
}
\source{
\url{https://www.lter.uaf.edu/data/data-detail/id/655}
}
\usage{
soilCores
}
\description{
Example unstandardized soil core dataset
}
\keyword{datasets}
